+++ cognite_toolkit/_cdf_tk/utils/_auxiliary.py
